% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_ds_helpers.R
\name{adjmats_from_info}
\alias{adjmats_from_info}
\title{adjmats_from_info}
\usage{
adjmats_from_info(info)
}
\arguments{
\item{info}{Either an edgeinfo or a coedgeinfo data structure}
}
\value{
Two adjacency matrices that are extracted from the info
structure, one for the cut edges and one for the fold edges.
}
\description{
adjmats_from_info
}
\examples{
info <- rbind(c(0,0,0,1,1,1,1,2,1),c(1,1,1,0,0,0,2,1,0))
adjmats_from_info(info)

}
