% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_ds_helpers.R
\name{coedgelist_from_he_ds}
\alias{coedgelist_from_he_ds}
\title{coedgelist_from_he_ds
Produce an extended edgelist from the information in a
half-edge data structure, but using the 'virtual' edges between faces
instead of the actual edges between vertices.}
\usage{
coedgelist_from_he_ds(he_ds)
}
\arguments{
\item{he_ds}{The half-edge data structure}
}
\value{
A 9 column matrix with information about the two endpoints
of of each edge, the labels of the endpoints, and an indication of
whether the edge is 'cut' or not.
}
\description{
coedgelist_from_he_ds
Produce an extended edgelist from the information in a
half-edge data structure, but using the 'virtual' edges between faces
instead of the actual edges between vertices.
}
