% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube_abstract_trees.R
\docType{data}
\name{cube_abstract_trees}
\alias{cube_abstract_trees}
\title{cube_abstract_trees}
\format{
An object of class \\code{array} of dimension 8 x 8 x 23.`
}
\source{
Generated by Sage: graphs.trees(8)
}
\usage{
cube_abstract_trees
}
\description{
The 23 trees with 8 vertices as an array of adjacency matrices
}
\keyword{datasets}
