% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube_automorphism_group.R
\docType{data}
\name{cube_automorphism_group}
\alias{cube_automorphism_group}
\title{cube_automorphism_group}
\format{
An object of class \\code{list} of length 4:
\describe{
\item{elements}{An object of class \\code{matrix} (inherits from \\code{array}) with 48 rows and 8 columns}
\item{names}{An object of class \\code{character} of length 48}
\item{perms}{An object of class \\code{character} of length 48}
\item{matrices}{An object of class \\code{array} of dimension 6 x 6 x 48}
}
}
\source{
Generated by GAP
}
\usage{
cube_automorphism_group
}
\description{
The 48 element automorphism group of the cube, with accompanying information.
The elements are ordered by conjugacy classes.
The names are derived from the conjugacy class names produced by GAP.
}
\keyword{datasets}
