% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube_span__trees.R
\docType{data}
\name{cube_span_trees}
\alias{cube_span_trees}
\title{cube_span_trees}
\format{
An object of class \\code{array} of dimension 8 x 8 x 384.
}
\source{
Generated by Sage: \\code{oct = Graph("GrQHGs")
			      oct.spanning_trees()}
}
\usage{
cube_span_trees
}
\description{
The 384 spanning trees ot the cube as an array of adjacency matrices
}
\keyword{datasets}
