% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_ds_helpers.R
\name{display_hes}
\alias{display_hes}
\title{display_hes
Display half-edge data structures with rgl.}
\usage{
display_hes(hes, rc, shared = FALSE, special = FALSE)
}
\arguments{
\item{hes}{A list of half-edge data structures}

\item{rc}{The number of rows and columns in the displayl r * c = length(hes)}

\item{shared}{Whether the rgl mouse interactions are shared or not}

\item{special}{a boolean indicating whether to show simplified view,
without cut and fold information.}
}
\value{
Nothing is returned from this function.
}
\description{
display_hes
Display half-edge data structures with rgl.
}
