% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_ds_helpers.R
\name{edgelist_to_adjmat}
\alias{edgelist_to_adjmat}
\title{edgelist_to_adjmat
Convert an edgelist to an adjacency matrix.
The edgelist is treated as an index array for the
desired matrix.}
\usage{
edgelist_to_adjmat(edgelist, dim = NULL)
}
\arguments{
\item{edgelist}{An n X 2 matrix of edges of a graph}

\item{dim}{Optional help for deciding the dimension of
the result; defaults to the largest value in the edgelist}
}
\value{
The adjacency matrix of the graph
}
\description{
edgelist_to_adjmat
Convert an edgelist to an adjacency matrix.
The edgelist is treated as an index array for the
desired matrix.
}
\examples{
elist <- rbind(c(1,4), c(2,4), c(3,4))
edgelist_to_adjmat(elist)

}
