% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/half_edge_data_structure.R
\name{Face}
\alias{Face}
\title{Face R6 Class}
\description{
Face R6 Class

Face R6 Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{ix}}{Index of this face.}

\item{\code{rot}}{Rotation of this face:
0 = 0 = 2\emph{pi, 1 = pi/2, 2 = pi, 3 = 3}pi/2.}

\item{\code{center}}{The center of this face: x,y,z coordinates.}

\item{\code{edge}}{The starting edge for traversing the perimeter of this face.
print
render an identifier for this object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Face-print}{\code{Face$print()}}
\item \href{#method-Face-clone}{\code{Face$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Face-print"></a>}}
\if{latex}{\out{\hypertarget{method-Face-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Face$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Face-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Face-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Face$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
