% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_ds_helpers.R
\name{get_display_info}
\alias{get_display_info}
\title{get_display_info
Process an edgelist to assemple the information into a
format amenible to rgl.}
\usage{
get_display_info(info)
}
\arguments{
\item{info}{Either an extended edgelist or a coedgelist}
}
\value{
A list of the extracted info
}
\description{
get_display_info
Process an edgelist to assemple the information into a
format amenible to rgl.
}
