% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/half_edge_data_structure.R
\name{HalfEdge}
\alias{HalfEdge}
\title{HalfEdge R6 Class}
\description{
HalfEdge R6 Class

HalfEdge R6 Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{ix}}{Index of this half-edge.}

\item{\code{dir}}{Direction this half-edge points}

\item{\code{vert}}{Vertex at the end of the half-edge}

\item{\code{pair}}{Oppositely oriented adjacent half-edge}

\item{\code{face}}{Face that the half-edge borders}

\item{\code{succ}}{Next half-edge around the face
print
render an identifier for this object}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{cut}}{Whether this half-edge is cut or not}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-HalfEdge-print}{\code{HalfEdge$print()}}
\item \href{#method-HalfEdge-clone}{\code{HalfEdge$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HalfEdge-print"></a>}}
\if{latex}{\out{\hypertarget{method-HalfEdge-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HalfEdge$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HalfEdge-clone"></a>}}
\if{latex}{\out{\hypertarget{method-HalfEdge-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HalfEdge$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
