% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_ds_helpers.R
\name{hull_plus_g6}
\alias{hull_plus_g6}
\title{hull_plus_g6
Treat the half-edge data structure as an arithmetic operator on
the indices of the spanning trees}
\usage{
hull_plus_g6(g6, hull)
}
\arguments{
\item{g6}{A spanning tree in g6 encoding}

\item{hull}{A hull structure for a polyhedron}
}
\value{
Vector of 4 adjacency matrices in g6 encoding.
Only 2 of them will be trees, and one of those trees will be
identical to the input tree.
}
\description{
hull_plus_g6
Treat the half-edge data structure as an arithmetic operator on
the indices of the spanning trees
}
