% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map12.R
\docType{data}
\name{map12}
\alias{map12}
\title{map12}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 12 rows and 7 columns.
}
\usage{
map12
}
\description{
A convenience mapping from abstract trees to spanning trees
}
\keyword{datasets}
