% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map372.R
\docType{data}
\name{map372}
\alias{map372}
\title{map372}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 372 rows and 7 columns.
}
\usage{
map372
}
\description{
A convenience mapping from abstract trees to spanning trees
}
\keyword{datasets}
