% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oct_abstract_trees.R
\docType{data}
\name{oct_abstract_trees}
\alias{oct_abstract_trees}
\title{oct_abstract_trees}
\format{
An object of class \\code{array} of dimension 6 x 6 x 6.
}
\source{
Generated by Sage: graphs.trees(6)
}
\usage{
oct_abstract_trees
}
\description{
The 6 trees with 6 vertices as an array of adjacency matrices
}
\keyword{datasets}
