% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oct_span__trees.R
\docType{data}
\name{oct_span_trees}
\alias{oct_span_trees}
\title{oct_span_trees}
\format{
An object of class \\code{array} of dimension 6 x 6 x 384.
}
\source{
Generated by Sage: oct = Graph("E\}lw")
oct.spanning_trees()
}
\usage{
oct_span_trees
}
\description{
The 384 spanning trees ot the octahedron as an array of adjacency matrices
}
\keyword{datasets}
