% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/octahedron_automorphism_group.R
\docType{data}
\name{octahedron_automorphism_group}
\alias{octahedron_automorphism_group}
\title{octahedron_automorphism_group}
\format{
An object of class \\code{list} of length 4:
\describe{
\item{elements}{An object of class \\code{matrix} (inherits from \\code{array}) with 48 rows and 8 columns}
\item{names}{An object of class \\code{character} of length 4}
\item{perms}{An object of class \\code{character} of length 48}
\item{matrices}{An object of class \\code{array} of dimension 6 x 6 x 48}
}
}
\source{
Generated by GAP
}
\usage{
octahedron_automorphism_group
}
\description{
The 48 element automorphism group of the octahedron, with accompanying information
}
\keyword{datasets}
