% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/half_edge_data_structure.R
\name{Vertex}
\alias{Vertex}
\title{Vertex R6 Class}
\description{
Vertex R6 Class

Vertex R6 Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{ix}}{Index of this vertex.}

\item{\code{point}}{The spatial location of this vertex: x,y,z coordinates.}

\item{\code{edge}}{The half-edge that starts at this vertex.}

\item{\code{pair}}{The perfect matching pair of this vertex.
print
render an identifier for this object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Vertex-print}{\code{Vertex$print()}}
\item \href{#method-Vertex-clone}{\code{Vertex$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Vertex-print"></a>}}
\if{latex}{\out{\hypertarget{method-Vertex-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Vertex$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Vertex-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Vertex-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Vertex$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
